refactor: remove translateToLedgerJournal from ledger/index.ts
